tss_api: refactor keygen v2
